Rewrite conform core framework to use dynamic scope instead of lexical.
This allows sub-conformlets to be instantiated with only a function call, decreasing the syntactic noise of the conform call.
Additionally, there is no longer a need for a distinction between configuration arguments and instance arguments -- instead,
they are both treated the same and to pass to a higher-order conformlet, simply curry the conformlet.